解决冲突

+++ hr-java/hr-domain/src/main/java/com/zeroone/star/project/query/orgmanager/DepartmentQuery.java
@@ -16,6 +16,6 @@ public class DepartmentQuery extends PageQuery {
     @ApiModelProperty(value = "页码", example = "1")
     private Integer page;
     @ApiModelProperty(value = "每页显示条数", example = "10")
-    private Integer pageSize;
+    private long pageSize;
 
 }
